Debuggg

+++ onmt/translate/beam.py
@@ -231,7 +231,7 @@ class Beam(object):
                 if self.next_ys[-1][i] == self._eos:
                     global_scores = self.global_scorer.score(self, self.scores)
                     s = global_scores[i]
-                    self.finished.append((s, len(self.next_ys) - 1, i))
+                    self.finished.append((s.double(), len(self.next_ys) - 1, i))
 
             # End condition is when top-of-beam is EOS and no global score.
             if self.next_ys[-1][0] == self._eos:
@@ -321,7 +321,7 @@ class Beam(object):
                 if self.next_ys[-1][i] == self._eos:
                     global_scores = self.global_scorer.score(self, self.scores)
                     s = global_scores[i]
-                    self.finished.append((s, len(self.next_ys) - 1, i))
+                    self.finished.append((s.double(), len(self.next_ys) - 1, i))
 
             # End condition is when top-of-beam is EOS and no global score.
             if self.next_ys[-1][0] == self._eos:
@@ -455,7 +455,7 @@ class Beam(object):
                 if self.next_ys[-1][i] == self._eos:
                     global_scores = self.global_scorer.score(self, self.scores)
                     s = global_scores[i]
-                    self.finished.append((s, len(self.next_ys) - 1, i))
+                    self.finished.append((s.double(), len(self.next_ys) - 1, i))
 
             # End condition is when top-of-beam is EOS and no global score.
             if self.next_ys[-1][0] == self._eos:
@@ -541,7 +541,7 @@ class Beam(object):
                 if self.next_ys[-1][i] == self._eos:
                     global_scores = self.global_scorer.score(self, self.scores)
                     s = global_scores[i]
-                    self.finished.append((s, len(self.next_ys) - 1, i))
+                    self.finished.append((s.double(), len(self.next_ys) - 1, i))
 
             # End condition is when top-of-beam is EOS and no global score.
             if self.next_ys[-1][0] == self._eos:
@@ -568,7 +568,7 @@ class Beam(object):
                 if self.next_ys[-1][i] == self._eos:
                     global_scores = self.global_scorer.score(self, self.scores)
                     s = global_scores[i]
-                    self.finished.append((s, len(self.next_ys) - 1, i))
+                    self.finished.append((s.double(), len(self.next_ys) - 1, i))
 
             # End condition is when top-of-beam is EOS and no global score.
             if self.next_ys[-1][0] == self._eos:
